+++ pydantic_resolve/__init__.py
@@ -11,7 +11,7 @@ from .util import (
     mapper,
     ensure_subset,
     output,
-    schema_config,
+    model_config,
     copy_dataloader_kls)
 
 
@@ -30,6 +30,6 @@ __all__ = [
     'mapper',
     'ensure_subset',
     'output',
-    'schema_config',
+    'model_config',
     'copy_dataloader_kls',
 ]