+++ tests/utils/test_schema_config.py
@@ -1,12 +1,12 @@
-from pydantic_resolve.util import schema_config
-from pydantic import BaseModel
+from pydantic_resolve.util import model_config
+from pydantic import BaseModel, Field
 from pydantic_resolve import Resolver
 import pytest
 
 @pytest.mark.asyncio
-async def test_schema_config():
+async def test_schema_config_hidden():
 
-    @schema_config(hidden_fields=['id', 'password2'], default_required=True)
+    @model_config(hidden_fields=['id', 'password'])
     class Y(BaseModel):
         id: int = 0
         def resolve_id(self):
@@ -14,8 +14,8 @@ async def test_schema_config():
 
         name: str
 
-        password2: str = ''
-        def resolve_password2(self):
+        password: str = ''
+        def resolve_password(self):
             return 'confidential'
 
     schema = Y.schema()
@@ -28,8 +28,34 @@ async def test_schema_config():
 
 
 @pytest.mark.asyncio
+async def test_schema_config_hidden_with_field():
+    """Field(exclude=True) will also work """
+
+    @model_config(hidden_fields=['id', 'password'])
+    class Y(BaseModel):
+        id: int = 0
+        def resolve_id(self):
+            return 1
+
+        name: str = Field(exclude=True)
+
+        password: str = ''
+        def resolve_password(self):
+            return 'confidential'
+
+    schema = Y.schema()
+    assert list((schema['properties']).keys()) == []
+
+    y = Y(name='kikodo')
+
+    y = await Resolver().resolve(y) 
+    assert y.dict() == {}
+
+
+
+@pytest.mark.asyncio
 async def test_schema_config_required():
-    @schema_config(hidden_fields=['password2'])
+    @model_config()
     class Y(BaseModel):
         id: int = 0
         def resolve_id(self):
@@ -37,17 +63,17 @@ async def test_schema_config_required():
 
         name: str
 
-        password2: str = ''
-        def resolve_password2(self):
+        password: str = ''
+        def resolve_password(self):
             return 'confidential'
 
     schema = Y.schema()
-    assert set(schema['required']) == {'id', 'name'}
+    assert set(schema['required']) == {'id', 'name', 'password'}
 
 
 @pytest.mark.asyncio
 async def test_schema_config_required_false():
-    @schema_config(default_required=False)
+    @model_config(default_required=False)
     class Y(BaseModel):
         id: int = 0
         def resolve_id(self):
@@ -55,8 +81,8 @@ async def test_schema_config_required_false():
 
         name: str
 
-        password2: str = ''
-        def resolve_password2(self):
+        password: str = ''
+        def resolve_password(self):
             return 'confidential'
 
     schema = Y.schema()